+++ satella/posix/signals.py
@@ -14,13 +14,16 @@ def __sighandler(a, b):
     end = True
 
 
-def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None):
+def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None,
+                   end_on_keyboard_interrupt: bool = True):
     """
     Will hang until this process receives SIGTERM or SIGINT.
     If you pass extra signal IDs (signal.SIG*) with extra_signals,
     then also on those signals this call will release.
 
     :param extra_signals: a list of extra signals to listen to
+    :param end_on_keyboard_interrupt: whether to consider receiving a KeyboardInterrupt as
+        a signal to finish
     """
     extra_signals = extra_signals or []
     global end
@@ -43,8 +46,6 @@ def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None):
             else:
                 bugged_pypy = year <= 2012
 
-    sleep = lambda: time.sleep(1) if bugged_pypy else signal.pause
-
     signal.signal(signal.SIGTERM, __sighandler)
     signal.signal(signal.SIGINT, __sighandler)
     for s in extra_signals:
@@ -52,11 +53,9 @@ def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None):
 
     while not end:
         try:
-            sleep()
-        except AttributeError:  # pause() is undefined on Windows
-            try:  # we will sleep for small periods of time
-                time.sleep(0.5)
-            except IOError:  # "Interrupted system call"
-                pass
+            time.sleep(0.5)
+        except KeyboardInterrupt:
+            if end_on_keyboard_interrupt:
+                end = True
 
     end = False  # reset for next use